pyproject.toml
[tool.poetry]
name = "tzinfo"
version = "0.1.0"
description = " display timezone info based on offset and keyword search"
authors = ["Rajkumar N <rajkumar.natarajan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.3"
requests = "^2.28.1"
PyYAML = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
